# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: RFP Agent API
  version: 1.0.0
  description: |
    **RFP Agent** is an AI-powered platform for automated RFP discovery, proposal generation, and submission management.

    ## Features
    - 🔍 **Automated RFP Discovery**: Scan government portals for new opportunities
    - 🤖 **AI-Powered Proposal Generation**: Generate compliant proposals using GPT-5
    - 📊 **3-Tier Agent System**: Orchestrator, Managers, and Specialist agents
    - 📄 **Document Intelligence**: Parse and analyze RFP documents
    - 🚀 **Automated Submission**: Submit proposals through portal automation
    - 📈 **Analytics & Monitoring**: Track performance and success rates

    ## Architecture
    The platform uses a sophisticated 3-tier agentic system:
    - **Tier 1**: Primary Orchestrator (1 agent)
    - **Tier 2**: Managers - Portal, Proposal, Research (3 agents)
    - **Tier 3**: Specialists - Scanner, Monitor, Generator, Checker, Processor, Market Analyst, Historical Analyzer (7 agents)

    ## Authentication
    Currently, the API uses session-based authentication. API key support is planned for future releases.

  contact:
    name: RFP Agent Support
    email: support@rfpagent.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://api.rfpagent.com/api
    description: Production server
...
paths:
...
  /rfps/{id}:
    get:
      tags: [RFPs]
      summary: Get RFP by ID
      description: Retrieve detailed information for a specific RFP
      operationId: getRFP
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: RFP ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RFP'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags: [RFPs]
      summary: Update an RFP
      description: Update RFP details
      operationId: updateRFP
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRFPRequest'
      responses:
        '200':
          description: RFP updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RFP'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [RFPs]
      summary: Delete an RFP
      description: Remove an RFP from the system
      operationId: deleteRFP
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: RFP deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    RFP:
      type: object
      required:
        - id
        - title
        - agency
        - sourceUrl
        - status
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        agency:
          type: string
        category:
          type: string
        portalId:
          type: string
          format: uuid
        sourceUrl:
          type: string
          format: uri
        deadline:
          type: string
          format: date-time
        estimatedValue:
          type: string
          pattern: '^\d+\.\d{2}$'
        status:
          type: string
          enum: [discovered, parsing, drafting, review, approved, submitted, closed]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        requirements:
          type: object
        complianceItems:
          type: object
        riskFlags:
          type: object
        analysis:
          type: object
        addedBy:
          type: string
          enum: [manual, automatic]
        discoveredAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateRFPRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [discovered, parsing, drafting, review, approved, submitted, closed]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        requirements:
          type: object
        complianceItems:
          type: object
        analysis:
          type: object
...
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: array
                items:
                  type: object
          example:
            error: "Invalid input data"
            details:
              - field: "url"
                message: "Must be a valid URL"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
          example:
            error: "Resource not found"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
          example:
            error: "Internal server error"

  # ============================================================================
  # SECURITY SCHEMES
  # ============================================================================
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session-based authentication using cookies

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication (planned for future release)

# Global security - currently using session-based auth
security:
  - sessionAuth: []